+++ language/spanish_colombia_utf-8.php
@@ -810,6 +810,7 @@ $LANG_SHOP = array (
 'dscp_wfadmin' => 'Manage order statuses. Indicate which statuses cause a buyer or admin notification and which statuses indicate accepted orders',
 'err_gw_version' => 'The uploaded gateway must support Shop v%s, v%s provided.',
 'err_occurred' => 'An error occurred, check the system log.',
+'order_pub_link' => 'Public link to view the order',
 );
 if (isset($_SHOP_CONF['ena_ratings']) && $_SHOP_CONF['ena_ratings']) {
     $LANG_SHOP['list_sort_options']['top_rated'] = 'Top Rated';